rename conn to sconn and backup app to app2
