--- app/app.go
@@ -1,1 +1,0 @@
-package app
