+++ app2/app.go
@@ -1,0 +1,1 @@
+package app2
