+++ app2/client/capture/capture_windows_test.go
