+++ app2/client/session/manager.go
