+++ app2/client/session/session.go
